+++ Changelog.md
@@ -9,8 +9,9 @@
 
 - When programming via MCUboot, Programmer now detects whether the selected
   `.zip` file is a modem firmware or an application.
-- When programming using MCUboot, the Programmer app now detects whether the selected
-  `.zip` file is a modem firmware or an application.
+- When programming using MCUboot, the Programmer app now detects whether the
+  selected `.zip` file is a modem firmware or an application.
+
 ## 4.7.5 - 2026-06-15
 
 ### Changed